requirements.txt
streamlit
pandas
Sqlite3
numpy
SQLAlchemy
datetime
yfinance
matplotlib.pyplot
altair